refactor: rename debug-tools -> dev
